requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bitarray==1.2.2
bleach==3.3.0
blinker==1.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cytoolz==0.11.0
decorator==4.4.2
defusedxml==0.7.1
ecdsa==0.16.1
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
gitdb==4.0.7
GitPython==3.1.14
hexbytes==0.2.1
idna==2.10
importlib-metadata==3.8.0
ipfshttpclient==0.7.0a1
ipykernel==5.5.0
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lru-dict==1.1.7
MarkupSafe==1.1.1
mistune==0.8.4
mnemonic==0.19
multiaddr==0.0.9
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
netaddr==0.8.0
notebook==6.3.0
numpy==1.20.1
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parsimonious==0.8.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.18
protobuf==3.15.6
ptyprocess==0.7.0
pyarrow==3.0.0
pycparser==2.20
pycryptodome==3.10.1
pydeck==0.6.1
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
pysha3==1.0.2
pytesseract==0.3.7
python-dateutil==2.8.1
python-dotenv==0.16.0
pytz==2021.1
pyzmq==22.0.3
requests==2.25.1
rlp==2.0.1
Send2Trash==1.5.0
six==1.15.0
smmap==4.0.0
streamlit==0.79.0
terminado==0.9.3
testpath==0.4.4
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
validators==0.18.2
varint==1.0.2
watchdog==2.0.2
wcwidth==0.2.5
web3==5.17.0
webencodings==0.5.1
websockets==8.1
widgetsnbextension==3.5.1
zipp==3.4.1
